+++ CHANGELOG.md
@@ -40,6 +40,7 @@ _Not yet on NuGet..._
 * Scatter: New `Add.ScatterLine()` method creates a scatter plot with a line only and no markers (#3462, #3452) @MCF
 * Scatter: New `Add.ScatterPoints()` method creates a scatter plot with markers only and no line (#3462, #3452) @MCF
 * Ticks: Improve performance by reducing the number of string measurements (#3468) @drolevar
+* Plot: `GetCoordinateRect()` now returns dimensions that respect `ScaleFactor` (#3471) @MCF
 
 ## ScottPlot 4.1.72
 _Not yet on NuGet..._